feat: pass test cases from backend
